[Gold II] Title: 사과와 바나나, Time: 696 ms, Memory: 213204 KB -BaekjoonHub

+++ 백준/Gold/3114. 사과와 바나나/사과와 바나나.java
@@ -29,7 +29,7 @@ public class Main {
             }
         }
 
-        long[][] dp = new long[R][C];
+        int[][] dp = new int[R][C];
         dp[0][0] = pa[R-1][0] - pa[0][0];
         // - \ |
